Move tab handling to a trait.
